MDEV-31361 Wrong result on second execution of prepared statement for engine Federated with derived table
The problem is caused by the logic which decides to set
the "mergeable" flag for a derived table before the pushdown handler
is chosen and the derived table (DT) is set for execution by an
external engine. In fact, if the DT is set for pushdown then it must
be only materialized and not merged.

This commit puts the logic of choosing a pushdown handler before
the logic of merge/materalization for DT, and also prevents setting
the "mergeable" flag for DTs which are going to be pushed down.

Also this commit renames method TABLE_LIST::init_derived() to
setup_derived(), because actually the method is called more than once
(for example, from mysql_derived_init() and from
mysql_derived_prepare()), but usually it is expected that something
may be "initialized" only once. The word "setup" does not bring such
an implicit meaning, so seems to be a better alternative here